fix(#531): extract genre normalizer to core module, test production export
CR finding: test suite called a copied helper so it couldn't detect
regressions in the parser implementation. Moved normalizeGenres to
src/main/core/normalizeGenres.ts, imported in both parseSong.ts and
the test file. Old getGenreInfoFromSong re-exported for compatibility.
